Boost free !!!

+++ source/VirtualFluidsCore/CoProcessors/WriteMacroscopicQuantitiesCoProcessor.cpp
@@ -176,11 +176,11 @@ void WriteMacroscopicQuantitiesCoProcessor::addDataMQ(Block3DPtr block)
    //D3Q27BoundaryConditionPtr bcPtr;
    int nr = (int)nodes.size();
  
-   for(size_t ix3=minX3; ix3<=maxX3; ix3++)
+   for(int ix3=minX3; ix3<=maxX3; ix3++)
    {
-      for(size_t ix2=minX2; ix2<=maxX2; ix2++)
+      for(int ix2=minX2; ix2<=maxX2; ix2++)
       {
-         for(size_t ix1=minX1; ix1<=maxX1; ix1++)
+         for(int ix1=minX1; ix1<=maxX1; ix1++)
          {
             if(!bcArray->isUndefined(ix1,ix2,ix3) && !bcArray->isSolid(ix1,ix2,ix3))
             {
